+++ tests/test_pbsa.py
@@ -1,8 +1,8 @@
 import sys
 from pathlib import Path
 
-from rush.client import RunOpts, save_json, set_opts
-from rush.pbsa import PBSAResults, pbsa
+from rush.client import RunError, RunOpts, set_opts
+from rush.pbsa import PBSAResult, fetch_outputs, pbsa, save_outputs
 
 
 def test_pbsa():
@@ -27,8 +27,11 @@ def test_pbsa():
         ),
         collect=True,
     )
-    print(PBSAResults(*res), file=sys.stderr)
-    save_json(res, name="test_pbsa.json")
+    output = fetch_outputs(res)
+    assert not isinstance(output, (str, RunError))
+    assert isinstance(output, PBSAResult)
+    print(output, file=sys.stderr)
+    save_outputs(res)
 
 
 if __name__ == "__main__":